Change ints in NS_ENUMs to NSIntegers.

+++ Sparkle/SUAutomaticUpdateAlert.h
@@ -13,7 +13,7 @@
 
 @protocol SUAutomaticUpdateAlertDelegate;
 
-typedef NS_ENUM(int, SUAutomaticInstallationChoice) {
+typedef NS_ENUM(NSInteger, SUAutomaticInstallationChoice) {
     SUInstallNowChoice,
     SUInstallLaterChoice,
     SUDoNotInstallChoice
